Near field manager (iOS): protect against dangling pointers
clearTarget calls 'invalidate' on a private target pimpl object,
which it's documented to own (via a public target object).
It's possible to call deleteLater on QNearFieldTarget object
but this potentially leaves us with a dandling pointer, since
we do not remove deleted object from 'detectedTargets'.
Protect against this using QPointer.

Change-Id: Ib970fde0233d383b727b90afeb2aedf7eb4a9e9c
Pick-to: 6.8
Fixes: QTBUG-129977
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>