Docs: clarify QLowEnergyAdvertisingData::setRawData() limitations
It is only available on Linux Kernel API backend, so explicitly mention
it in the docs. Also explain, why it cannot be supported on other
backends with the current API.

As the new Linux D-Bus backend, together with the Linux Specific docs
section, was introduced only in 6.5, pick this patch down to that
branch only.

Task-number: QTBUG-130146
Task-number: QTBUG-130891
Pick-to: 6.8 6.5
Change-Id: I5395fed6606d84ed1d1f032c45b31759330293dd
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>