Docs: clarify that Apple does not support manufacturer data advertising
The Apple docs for startAdvertising [0] explicitly mention that it
supports only two advertising keys - for the local name and UUIDs.

I verified that CBAdvertisementDataManufacturerDataKey is simply
ignored. According to a comment in our code, such limitation exists
at least since 2016. Let's be honest with our users and document this
limitation.

[0]: https://developer.apple.com/documentation/corebluetooth/cbperipheralmanager/startadvertising(_:)?language=objc

Task-number: QTBUG-130146
Pick-to: 6.8 6.5
Change-Id: I36892672cbef76f4363487c1060ec38649d1018a
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>