Suppress characteristic/descriptor get/setValue() deprecation warnings
These two classes handle the deprecation of the methods and suppressing
the warning is sufficient.

For clarity it should be mentioned that these classes are used on
the peripheral (server) side, where local caching is needed. The local
caching directly into the Android's descriptor/characteristic
objects themselves was deprecated in API level 33 (with deprecation
of setValue()/getValue()).

Pick-to: 6.9 6.8 6.5
Task-number: QTBUG-132455
Change-Id: I71e6f584a09e7a98c9c17f5f1a2765183d5437f5
Reviewed-by: Lars Schmertmann <lars.schmertmann@governikus.de>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>