BlueZ: rename generated DBus propperites interface class
To avoid symbol clashing with the same interface added to QtGui in
qtbase/38251c36edf11316a2467169b1d491bf13520fd3.
Conveniently, qdbusxml2cpp already provides a command line option to
set a custom name for the generated class.

It's easier to fix it on QtBluetooth side, because the files are
generated by the script and later committed to the repo, while on
QtGui side they are handled directly by CMake, which for now provides
less granularity.

As a drive-by: fix an error in the DBus interface description.

Fixes: QTBUG-133553
Pick-to: 6.9 6.8
Change-Id: I2ab413bf617f53e26c37e430198e58071c84d49c
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>